extract resolver config discovery code
