+++ packages/compat/tests/audit.test.ts
@@ -62,7 +62,6 @@ describe('audit', function () {
         },
       ],
       resolvableExtensions,
-      autoRun: true,
       staticAppPaths: [],
       emberVersion: '4.0.0',
     };
