+++ packages/core/src/index.ts
@@ -13,7 +13,8 @@ export { default as toBroccoliPlugin } from './to-broccoli-plugin';
 export { default as WaitForTrees, OutputPaths } from './wait-for-trees';
 export { compile as jsHandlebarsCompile } from './js-handlebars';
 export { todo, unsupported, warn, debug, expectWarning, throwOnWarnings } from './messages';
-export { Resolver, Options as ResolverOptions, ModuleRequest, Resolution } from './module-resolver';
+export { Resolver, ModuleRequest, Resolution } from './module-resolver';
+export type { Options as ResolverOptions } from './module-resolver-options';
 export { ResolverLoader } from './resolver-loader';
 export { virtualContent } from './virtual-content';
 export type { Engine } from './app-files';
