+++ tests/scenarios/compat-resolver-test.ts
@@ -97,7 +97,6 @@ Scenarios.fromProject(() => new Project())
                 ...extraOpts?.appPackageRules,
               },
             ],
-            autoRun: true,
             staticAppPaths: [],
             emberVersion,
           };
