+++ tests/scenarios/core-resolver-test.ts
@@ -135,7 +135,6 @@ Scenarios.fromProject(() => new Project())
                 roots: [app.dir],
               },
             ],
-            autoRun: true,
             staticAppPaths: [],
             emberVersion: '4.0.0',
           };
